Moves parser module to parser/parsemd
This prepares the codebase to receive a new parser (ie parseorg) by removing
the assumption that "parser" means "Markdown parser."
